requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.6
aiosignal==1.3.1
aiosmtplib==3.0.2
aiosqlite==0.17.0
annotated-types==0.7.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
frozenlist==1.4.1
idna==3.10
iso8601==1.1.0
multidict==6.1.0
pydantic==2.9.2
pydantic_core==2.23.4
pypika-tortoise==0.1.6
python-dotenv==1.0.1
pytz==2024.2
schedule==1.2.2
tortoise-orm==0.21.6
typing_extensions==4.12.2
yarl==1.13.0
